pyproject.toml
[build-system]
requires = [ "setuptools >= 75.0" ]
build-backend = "setuptools.build_meta"

[project]
name = "data_formulator"
version = "0.7.0a2"

requires-python = ">=3.11"
authors = [
  {name = "Chenglong Wang", email = "chenglong.wang@microsoft.com"},
  {name = "Dan Marshall", email = "danmar@microsoft.com"},
]
readme = "README.md"
license = "MIT"
description = "Data Formulator is research protoype data visualization tool powered by AI."
keywords = ["data visualization", "LLM", "AI"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

dependencies = [
    "jupyter",
    "pandas",
    "flask",
    "flask-cors",
    "flask-limiter",
    "openai",
    "python-dotenv",
    "vega_datasets",
    "litellm",
    "duckdb",
    "numpy",
    "vl-convert-python",
    "backoff",
    "beautifulsoup4",
    "scikit-learn",
    "pyyaml",
    "pyarrow>=13.0.0",
    "xlrd",
    "openpyxl>=3.1.0",
    "yfinance",
    # Data-loader deps (all included by default; try/except keeps things safe)
    "pymongo",                  # mongodb
    "azure-cosmos",             # cosmosdb
    "azure-storage-blob",       # azure_blob
    "azure-identity",           # azure_blob, kusto
    "azure-kusto-data",         # kusto
    "azure-keyvault-secrets",   # kusto
    "pymysql",                  # mysql
    "pyodbc",                   # mssql
    "psycopg2-binary",          # postgresql
    "boto3",                    # s3, athena
    "google-cloud-bigquery",    # bigquery
    "google-auth",              # bigquery
    "db-dtypes",                # bigquery
    # SSO / Auth deps
    "PyJWT[crypto]>=2.8.0",     # OIDC JWT verification (includes cryptography)
    "requests",                 # GitHub OAuth code exchange, Superset API calls
    "flask-session>=0.8.0",     # Server-side session (SQLite) for TokenStore
]

[project.urls]
Homepage = "https://github.com/microsoft/data-formulator"
Repository = "https://github.com/microsoft/data-formulator.git"
"Bug Tracker" = "https://github.com/microsoft/data-formulator/issues"

[tool.setuptools]
package-dir = {"" = "py-src"}
include-package-data = true

[project.scripts]
data_formulator = "data_formulator:run_app"

[tool.uv]
dev-dependencies = [
    "build",
    "pytest",
]
